+++ graph/research_data/ranking-down.md
@@ -1,9 +1,9 @@
 |         | 比較したバージョン        | 環境部のアーティクルの種類の数 | byで引用するtheorem, label | proofの数 | 
 |---------|-------------------------|------------------------------|---------------------------|-----------| 
 | grcat_1 | 2012-12-02 and 2020-06-18 | +0, -0 | +0, -2 | -2 | 
-| monoid_0 | 2012-12-02 and 2020-06-18 | +0, -1 | +0, -12 | -25 | 
+| monoid_0 | 2012-12-02 and 2020-06-18 | +0, -1 | +0, -11 | -25 | 
 | pzfmisc1 | 2012-12-02 and 2020-06-18 | +0, -0 | +0, -6 | 0 | 
-| revrot_1 | 2012-12-02 and 2020-06-18 | +0, -1 | +8, -51 | -26 | 
+| revrot_1 | 2012-12-02 and 2020-06-18 | +0, -1 | +8, -50 | -26 | 
 | facirc_1 | 2012-12-02 and 2020-06-18 | +1, -0 | +0, -5 | -6 | 
 | limfunc1 | 2013-01-29 and 2020-06-18 | +0, -0 | +1, -0 | 0 | 
 | circcomb | 2012-12-02 and 2020-06-18 | +0, -0 | +0, -8 | -4 | 
@@ -11,32 +11,32 @@
 | borsuk_1 | 2012-12-02 and 2020-06-18 | +0, -0 | +0, -5 | 0 | 
 | jordan1 | 2012-12-02 and 2020-06-18 | +0, -0 | +0, -6 | -2 | 
 | pre_circ | 2013-12-10 and 2020-06-18 | +0, -0 | +0, -1 | 0 | 
-| rsspace3 | 2012-03-05 and 2020-06-18 | +2, -0 | +6, -12 | -3 | 
+| rsspace3 | 2012-03-05 and 2020-06-18 | +2, -0 | +5, -10 | -3 | 
 | gcd_1 | 2012-12-02 and 2020-06-18 | +0, -1 | +0, -13 | -4 | 
 | topmetr3 | 2012-03-05 and 2020-06-18 | +0, -0 | +5, -2 | 0 | 
 | cqc_the1 | 2012-03-05 and 2020-06-18 | +0, -0 | +9, -24 | -28 | 
 | moebius1 | 2012-03-05 and 2020-06-18 | +0, -1 | +1, -6 | -4 | 
-| waybel26 | 2012-03-05 and 2019-03-29 | +3, -0 | +3, -7 | -2 | 
+| waybel26 | 2012-03-05 and 2019-03-29 | +3, -0 | +2, -6 | -2 | 
 | nattra_1 | 2012-12-02 and 2020-06-18 | +0, -0 | +0, -11 | -1 | 
-| ec_pf_1 | 2012-12-02 and 2020-06-18 | +0, -0 | +1, -21 | -4 | 
-| sin_cos | 2013-12-10 and 2020-06-18 | +0, -0 | +0, -4 | 0 | 
-| extreal2 | 2012-03-05 and 2013-01-29 | +4, -3 | +2, -4 | 1 | 
+| ec_pf_1 | 2012-12-02 and 2020-06-18 | +0, -0 | +1, -19 | -4 | 
+| sin_cos | 2013-12-10 and 2020-06-18 | +0, -0 | +0, -2 | 0 | 
+| extreal2 | 2012-03-05 and 2013-01-29 | +4, -3 | +1, -4 | 1 | 
 | complfld | 2012-12-02 and 2020-06-18 | +2, -0 | +1, -5 | -7 | 
 | pralg_2 | 2012-12-02 and 2017-11-05 | +0, -1 | +0, -3 | 0 | 
 | vectsp_4 | 2012-12-02 and 2020-06-18 | +0, -0 | +1, -4 | -1 | 
 | altcat_1 | 2012-12-02 and 2020-06-18 | +0, -0 | +0, -8 | -2 | 
 | compos_1 | 2012-12-02 and 2020-06-18 | +0, -1 | +19, -17 | 4 | 
 | bhsp_4 | 2012-12-02 and 2020-06-18 | +0, -0 | +0, -4 | -4 | 
-| yellow10 | 2012-03-05 and 2020-06-18 | +0, -0 | +2, -7 | -2 | 
-| nat_lat | 2012-12-02 and 2020-06-18 | +0, -0 | +0, -9 | -2 | 
+| yellow10 | 2012-03-05 and 2020-06-18 | +0, -0 | +0, -5 | -2 | 
+| nat_lat | 2012-12-02 and 2020-06-18 | +0, -0 | +0, -5 | -2 | 
 | wellset1 | 2013-12-10 and 2020-06-18 | +0, -0 | +0, -0 | 0 | 
 | topreal3 | 2012-12-02 and 2016-07-24 | +0, -0 | +0, -8 | 0 | 
 | mod_4 | 2013-01-29 and 2015-10-30 | +0, -1 | +1, -12 | -11 | 
-| lopban_2 | 2012-12-02 and 2020-06-18 | +0, -0 | +1, -9 | -1 | 
+| lopban_2 | 2012-12-02 and 2020-06-18 | +0, -0 | +1, -8 | -1 | 
 | measur10 | 2016-07-24 and 2020-06-18 | +0, -0 | +3, -2 | 2 | 
 | relset_2 | 2012-03-05 and 2015-10-30 | +0, -0 | +1, -3 | -2 | 
 | polynom1 | 2012-12-02 and 2017-11-05 | +0, -0 | +0, -8 | -3 | 
-| rsspace2 | 2012-03-05 and 2012-12-02 | +0, -0 | +3, -5 | 1 | 
+| rsspace2 | 2012-03-05 and 2012-12-02 | +0, -0 | +2, -4 | 1 | 
 | yellow13 | 2012-12-02 and 2020-06-18 | +0, -0 | +0, -8 | -1 | 
 | gtarski2 | 2016-07-24 and 2020-06-18 | +0, -0 | +0, -0 | 0 | 
 | clvect_2 | 2012-03-05 and 2020-06-18 | +0, -1 | +0, -6 | -1 | 
@@ -46,40 +46,40 @@
 | memstr_0 | 2013-01-29 and 2020-06-18 | +0, -0 | +1, -2 | -5 | 
 | lattad_1 | 2015-02-12 and 2020-06-18 | +0, -0 | +0, -0 | 0 | 
 | msafree2 | 2012-12-02 and 2020-06-18 | +0, -0 | +0, -8 | -2 | 
-| topreal9 | 2012-03-05 and 2016-07-24 | +2, -2 | +42, -36 | 2 | 
+| topreal9 | 2012-03-05 and 2016-07-24 | +2, -2 | +41, -35 | 2 | 
 | graph_4 | 2013-12-10 and 2020-06-18 | +1, -0 | +17, -18 | 0 | 
 | euclid_3 | 2012-03-05 and 2020-06-18 | +1, -2 | +17, -19 | 0 | 
 | hilbert2 | 2012-03-05 and 2020-06-18 | +1, -0 | +1, -7 | -2 | 
-| kolmog01 | 2012-03-05 and 2020-06-18 | +1, -0 | +1, -8 | -2 | 
+| kolmog01 | 2012-03-05 and 2020-06-18 | +1, -0 | +1, -7 | -2 | 
 | goboard8 | 2012-12-02 and 2020-06-18 | +0, -0 | +0, -0 | 0 | 
 | pua2mss1 | 2012-12-02 and 2018-04-06 | +0, -0 | +0, -18 | 0 | 
-| brouwer2 | 2012-03-05 and 2020-06-18 | +3, -0 | +2, -18 | 0 | 
+| brouwer2 | 2012-03-05 and 2020-06-18 | +3, -0 | +2, -17 | 0 | 
 | msualg_7 | 2012-12-02 and 2020-06-18 | +0, -0 | +0, -8 | 0 | 
 | polyeq_2 | 2012-03-05 and 2020-06-18 | +0, -0 | +0, -2 | 0 | 
-| finance3 | 2016-07-24 and 2020-06-18 | +1, -0 | +4, -15 | -37 | 
+| finance3 | 2016-07-24 and 2020-06-18 | +1, -0 | +3, -13 | -37 | 
 | waybel_0 | 2013-12-10 and 2020-06-18 | +0, -0 | +0, -0 | 0 | 
 | zmodul06 | 2015-02-12 and 2020-06-18 | +7, -0 | +29, -14 | 0 | 
 | msaterm | 2012-12-02 and 2020-06-18 | +0, -0 | +1, -4 | 0 | 
 | lopban_5 | 2012-03-05 and 2015-10-30 | +1, -0 | +7, -1 | 0 | 
-| csspace | 2012-03-05 and 2019-03-29 | +3, -2 | +21, -35 | -27 | 
-| jordan19 | 2012-03-05 and 2020-06-18 | +2, -2 | +8, -9 | 0 | 
+| csspace | 2012-03-05 and 2019-03-29 | +3, -2 | +19, -35 | -27 | 
+| jordan19 | 2012-03-05 and 2020-06-18 | +2, -2 | +8, -8 | 0 | 
 | normsp_1 | 2013-12-10 and 2018-04-06 | +0, -0 | +0, -0 | 0 | 
 | orders_3 | 2012-12-02 and 2020-06-18 | +0, -0 | +0, -9 | -3 | 
-| prvect_2 | 2012-03-05 and 2020-06-18 | +0, -0 | +8, -15 | -8 | 
+| prvect_2 | 2012-03-05 and 2020-06-18 | +0, -0 | +8, -12 | -8 | 
 | extens_1 | 2012-12-02 and 2020-06-18 | +0, -0 | +0, -6 | -1 | 
 | ring_3 | 2015-10-30 and 2020-06-18 | +0, -0 | +0, -4 | 0 | 
-| aofa_a01 | 2012-12-02 and 2020-06-18 | +0, -0 | +10, -25 | -6 | 
-| csspace2 | 2012-03-05 and 2020-06-18 | +3, -1 | +13, -19 | -3 | 
+| aofa_a01 | 2012-12-02 and 2020-06-18 | +0, -0 | +10, -24 | -6 | 
+| csspace2 | 2012-03-05 and 2020-06-18 | +3, -1 | +9, -15 | -3 | 
 | asympt_0 | 2012-03-05 and 2020-06-18 | +0, -0 | +5, -7 | 2 | 
-| borsuk_7 | 2012-03-05 and 2019-03-29 | +2, -2 | +31, -52 | -10 | 
+| borsuk_7 | 2012-03-05 and 2019-03-29 | +2, -2 | +30, -52 | -10 | 
 | groupp_1 | 2012-03-05 and 2020-06-18 | +1, -2 | +0, -4 | -3 | 
 | gtarski1 | 2015-02-12 and 2020-06-18 | +0, -0 | +0, -0 | 0 | 
 | euclid11 | 2015-10-30 and 2020-06-18 | +0, -0 | +0, -0 | 0 | 
 | yellow_6 | 2012-12-02 and 2020-06-18 | +0, -0 | +1, -16 | -10 | 
 | sin_cos5 | 2012-03-05 and 2020-06-18 | +1, -1 | +0, -1 | 0 | 
-| matrtop3 | 2012-03-05 and 2020-06-18 | +3, -3 | +8, -15 | -5 | 
+| matrtop3 | 2012-03-05 and 2020-06-18 | +3, -3 | +8, -14 | -5 | 
 | euclid_5 | 2012-03-05 and 2020-06-18 | +3, -1 | +5, -4 | 1 | 
-| mfold_0 | 2015-02-12 and 2020-06-18 | +2, -1 | +24, -21 | 4 | 
+| mfold_0 | 2015-02-12 and 2020-06-18 | +2, -1 | +22, -21 | 4 | 
 | transgeo | 2012-12-02 and 2020-06-18 | +0, -0 | +0, -11 | -8 | 
 | group_19 | 2015-02-12 and 2020-06-18 | +1, -0 | +5, -2 | -2 | 
 | binari_4 | 2012-03-05 and 2020-06-18 | +2, -1 | +1, -0 | -1 | 
@@ -96,10 +96,10 @@
 | euler_2 | 2013-12-10 and 2020-06-18 | +0, -0 | +0, -1 | 0 | 
 | rlaffin1 | 2012-03-05 and 2013-01-29 | +1, -3 | +5, -15 | 0 | 
 | euclid10 | 2015-02-12 and 2019-03-29 | +0, -0 | +0, -0 | 0 | 
-| toprealc | 2012-03-05 and 2016-07-24 | +1, -1 | +3, -8 | -3 | 
+| toprealc | 2012-03-05 and 2016-07-24 | +1, -1 | +2, -7 | -3 | 
 | compos_2 | 2012-03-05 and 2020-06-18 | +0, -1 | +20, -31 | 1 | 
 | bcialg_3 | 2012-12-02 and 2020-06-18 | +0, -0 | +0, -4 | -9 | 
 | jgraph_7 | 2012-03-05 and 2020-06-18 | +0, -0 | +0, -1 | 0 | 
-| vectmetr | 2012-03-05 and 2020-06-18 | +2, -1 | +3, -9 | 2 | 
-| glib_000 | 2012-03-05 and 2020-06-18 | +1, -0 | +12, -21 | 7 | 
-| AVERAGE | ***** | 0.34 | 13.76 | -2.77
+| vectmetr | 2012-03-05 and 2020-06-18 | +2, -1 | +3, -7 | 2 | 
+| glib_000 | 2012-03-05 and 2020-06-18 | +1, -0 | +12, -20 | 7 | 
+| AVERAGE | ***** | 0.34 | 13.22 | -2.77
