+++ graph/research_data/ranking-up.md
@@ -1,11 +1,11 @@
 |         | 比較したバージョン        | 環境部のアーティクルの種類の数 | byで引用するtheorem, label | proofの数 | 
 |---------|-------------------------|------------------------------|---------------------------|-----------| 
-| card_5 | 2012-03-05 and 2013-12-10 | +2, -0 | +8, -17 | -4 | 
-| measure6 | 2012-03-05 and 2013-01-29 | +0, -0 | +4, -13 | 0 | 
+| card_5 | 2012-03-05 and 2013-12-10 | +2, -0 | +8, -16 | -4 | 
+| measure6 | 2012-03-05 and 2013-01-29 | +0, -0 | +3, -11 | 0 | 
 | nat_2 | 2012-03-05 and 2013-12-10 | +0, -0 | +0, -0 | -1 | 
 | limfunc1 | 2012-03-05 and 2013-01-29 | +0, -0 | +6, -13 | 9 | 
 | finseq_6 | 2012-03-05 and 2013-12-10 | +0, -0 | +5, -9 | 0 | 
-| scmfsa_2 | 2012-12-02 and 2016-07-24 | +1, -0 | +0, -9 | -3 | 
+| scmfsa_2 | 2012-12-02 and 2016-07-24 | +1, -0 | +0, -8 | -3 | 
 | pre_circ | 2012-03-05 and 2013-12-10 | +0, -1 | +8, -8 | -2 | 
 | matrix_1 | 2013-01-29 and 2013-12-10 | +11, -0 | +29, -0 | 35 | 
 | graph_2 | 2012-03-05 and 2013-01-29 | +0, -1 | +2, -13 | 0 | 
@@ -16,12 +16,12 @@
 | normsp_1 | 2012-12-02 and 2013-12-10 | +0, -0 | +0, -2 | -1 | 
 | card_fil | 2012-03-05 and 2015-02-12 | +0, -0 | +1, -17 | -8 | 
 | revrot_1 | 2012-03-05 and 2012-12-02 | +0, -2 | +1, -2 | 0 | 
-| msscyc_1 | 2012-03-05 and 2013-01-29 | +0, -0 | +5, -22 | -1 | 
-| ordinal6 | 2012-12-02 and 2016-07-24 | +1, -3 | +1, -11 | -8 | 
+| msscyc_1 | 2012-03-05 and 2013-01-29 | +0, -0 | +5, -21 | -1 | 
+| ordinal6 | 2012-12-02 and 2016-07-24 | +1, -3 | +0, -11 | -8 | 
 | scmfsa6a | 2012-12-02 and 2019-03-29 | +2, -1 | +20, -7 | 6 | 
 | fdiff_1 | 2013-01-29 and 2015-02-12 | +0, -0 | +0, -2 | -1 | 
 | wsierp_1 | 2012-03-05 and 2013-12-10 | +0, -3 | +2, -1 | -1 | 
-| orders_2 | 2012-03-05 and 2013-01-29 | +1, -1 | +0, -28 | -2 | 
+| orders_2 | 2012-03-05 and 2013-01-29 | +1, -1 | +1, -27 | -2 | 
 | newton04 | 2019-03-29 and 2020-06-18 | +0, -0 | +1, -3 | -1 | 
 | fuzzy_2 | 2013-01-29 and 2018-05-15 | +0, -0 | +0, -0 | 0 | 
 | hilbasis | 2017-11-05 and 2018-04-06 | +0, -0 | +0, -0 | 0 | 
@@ -41,9 +41,9 @@
 | polyeq_3 | 2012-03-05 and 2016-07-24 | +2, -3 | +2, -3 | 0 | 
 | jordan1 | 2012-03-05 and 2012-12-02 | +1, -0 | +7, -8 | 0 | 
 | toler_1 | 2012-03-05 and 2013-12-10 | +1, -0 | +1, -14 | -9 | 
-| fomodel2 | 2012-03-05 and 2016-07-24 | +3, -1 | +10, -25 | -19 | 
+| fomodel2 | 2012-03-05 and 2016-07-24 | +3, -1 | +10, -24 | -19 | 
 | borsuk_7 | 2019-03-29 and 2020-06-18 | +0, -0 | +0, -0 | 0 | 
-| ratfunc1 | 2013-01-29 and 2015-02-12 | +0, -0 | +2, -7 | -1 | 
+| ratfunc1 | 2013-01-29 and 2015-02-12 | +0, -0 | +2, -6 | -1 | 
 | prvect_3 | 2013-01-29 and 2016-07-24 | +0, -0 | +0, -0 | 0 | 
 | rinfsup1 | 2013-01-29 and 2015-02-12 | +0, -0 | +1, -0 | -4 | 
 | nattra_1 | 2012-03-05 and 2012-12-02 | +0, -3 | +52, -49 | 0 | 
@@ -55,7 +55,7 @@
 | sin_cos2 | 2012-03-05 and 2013-12-10 | +0, -0 | +1, -0 | -1 | 
 | extreal1 | 2012-03-05 and 2012-12-02 | +2, -3 | +4, -7 | -2 | 
 | incsp_1 | 2016-07-24 and 2017-11-05 | +0, -0 | +0, -0 | 0 | 
-| compos_1 | 2012-03-05 and 2012-12-02 | +0, -3 | +22, -24 | 0 | 
+| compos_1 | 2012-03-05 and 2012-12-02 | +0, -3 | +22, -23 | 0 | 
 | graph_4 | 2012-03-05 and 2013-12-10 | +0, -0 | +1, -5 | -5 | 
 | numpoly1 | 2015-02-12 and 2020-06-18 | +0, -0 | +0, -1 | 0 | 
 | tietze_2 | 2016-07-24 and 2017-11-05 | +0, -0 | +0, -0 | 0 | 
@@ -66,17 +66,17 @@
 | memstr_0 | 2012-03-05 and 2013-01-29 | +0, -5 | +0, -1 | -4 | 
 | altcat_1 | 2012-03-05 and 2012-12-02 | +0, -0 | +8, -9 | -1 | 
 | polynom3 | 2012-03-05 and 2015-02-12 | +1, -0 | +9, -13 | -5 | 
-| autalg_1 | 2012-03-05 and 2013-01-29 | +0, -0 | +0, -9 | -1 | 
+| autalg_1 | 2012-03-05 and 2013-01-29 | +0, -0 | +0, -8 | -1 | 
 | gaussint | 2016-07-24 and 2020-06-18 | +0, -0 | +0, -1 | -2 | 
-| sfmastr2 | 2012-12-02 and 2016-07-24 | +4, -0 | +6, -32 | -7 | 
+| sfmastr2 | 2012-12-02 and 2016-07-24 | +4, -0 | +6, -30 | -7 | 
 | hurwitz2 | 2018-05-15 and 2019-03-29 | +0, -0 | +0, -0 | 0 | 
 | csspace3 | 2016-07-24 and 2017-11-05 | +0, -0 | +0, -0 | 0 | 
 | euclid_9 | 2015-02-12 and 2015-10-30 | +0, -0 | +0, -0 | 0 | 
-| ndiff_5 | 2012-03-05 and 2015-02-12 | +0, -1 | +35, -15 | 11 | 
-| polyform | 2012-03-05 and 2019-03-29 | +1, -1 | +7, -24 | -4 | 
+| ndiff_5 | 2012-03-05 and 2015-02-12 | +0, -1 | +33, -13 | 11 | 
+| polyform | 2012-03-05 and 2019-03-29 | +1, -1 | +6, -23 | -4 | 
 | neckla_3 | 2015-02-12 and 2020-06-18 | +0, -0 | +0, -0 | 0 | 
 | clvect_1 | 2012-03-05 and 2018-04-06 | +0, -2 | +2, -7 | -10 | 
-| rsspace2 | 2012-12-02 and 2015-10-30 | +1, -0 | +7, -7 | -2 | 
+| rsspace2 | 2012-12-02 and 2015-10-30 | +1, -0 | +6, -6 | -2 | 
 | osalg_1 | 2015-02-12 and 2016-07-24 | +0, -0 | +0, -0 | 0 | 
 | sysrel | 2012-03-05 and 2013-01-29 | +0, -0 | +0, -6 | -3 | 
 | trees_a | 2012-03-05 and 2013-12-10 | +2, -0 | +1, -6 | 0 | 
@@ -85,12 +85,12 @@
 | toprealc | 2016-07-24 and 2017-11-05 | +0, -0 | +0, -1 | -1 | 
 | nat_5 | 2013-01-29 and 2018-04-06 | +0, -0 | +2, -6 | 0 | 
 | moebius2 | 2016-07-24 and 2020-06-18 | +0, -0 | +1, -4 | 0 | 
-| nat_6 | 2015-02-12 and 2020-06-18 | +1, -0 | +71, -117 | -2 | 
+| nat_6 | 2015-02-12 and 2020-06-18 | +1, -0 | +71, -119 | -2 | 
 | group_17 | 2016-07-24 and 2018-05-15 | +0, -0 | +0, -0 | 0 | 
 | circuit1 | 2012-03-05 and 2013-12-10 | +1, -0 | +2, -2 | 0 | 
 | radix_2 | 2012-03-05 and 2016-07-24 | +0, -2 | +1, -1 | 2 | 
 | lang1 | 2012-03-05 and 2013-12-10 | +1, -0 | +0, -3 | -1 | 
-| aofa_a00 | 2012-12-02 and 2016-07-24 | +1, -0 | +32, -42 | -31 | 
+| aofa_a00 | 2012-12-02 and 2016-07-24 | +1, -0 | +32, -41 | -31 | 
 | integra3 | 2012-03-05 and 2013-12-10 | +1, -0 | +1, -6 | -4 | 
 | rlaffin3 | 2013-01-29 and 2018-04-06 | +0, -0 | +0, -0 | 0 | 
 | rlaffin1 | 2013-01-29 and 2017-11-05 | +0, -0 | +0, -8 | -3 | 
@@ -102,4 +102,4 @@
 | integra9 | 2016-07-24 and 2019-03-29 | +0, -0 | +0, -0 | 0 | 
 | algstr_4 | 2013-01-29 and 2018-04-06 | +0, -0 | +0, -4 | 0 | 
 | matrtop1 | 2015-02-12 and 2017-11-05 | +0, -0 | +1, -0 | 0 | 
-| AVERAGE | ***** | 0.05 | 12.08 | -1.37
+| AVERAGE | ***** | 0.05 | 11.88 | -1.37
